requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
slack-sdk==3.26.1
openai==1.3.7
httpx==0.25.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
aiofiles==23.2.1
redis==5.0.1
